pyproject.toml
[project]
name = "agnoproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.0.10",
    "ddgs>=9.6.0",
    "google-genai>=1.39.0",
    "google-generativeai>=0.8.5",
    "lxml>=6.0.2",
    "python-dotenv>=1.1.1",
]
